split customEventDispatcher into separate file. add unit test for setLocale
